+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 <details>
 
 <summary>Released TODO</summary>
+
+* feature: add project composition using `packages` keyword in template file (see README)
 * feature: `GroupByWithRankOperation` cumulatively sums record counts by group-by columns
 * feature: setting `log_level: DEBUG` in template configs or setting `debug: True` for a node displays the head of the node mid-run 
 * feature: add `optional_fields` key to all Sources to add optional empty columns when missing from schema
@@ -10,6 +12,10 @@
 * internal: force-cast a dataframe to string-type before writing as a Destination
 * internal: remove attempted directory-hashing when a source is a directory (i.e., Parquet)
 * internal: refactor project to standardize import paths for Node and Operation
+* internal: add `Node.full_name` attribute and `Node.set_upstream_source()` method
+* internal: unify graph-building into compilation
+* internal: refactor compilation and execution code for cleanliness
+* internal: unify `Node.compile()` into initialization to ease Node development
 
 </details>